Add support for bigboy inputs

+++ src/aoc_2015/day01.rs
@@ -5,18 +5,18 @@ pub fn run(input: &str) {
     println!("Floor: {}, Position: {}", floor, basement_position);
 }
 
-fn get_floor(input: &str) -> i32 {
+fn get_floor(input: &str) -> i64 {
     let up_token = '(';
     let down_token = ')';
 
-    let down_count: i32 = input.matches(down_token).count() as i32;
-    let up_count: i32 = input.matches(up_token).count() as i32;
-    let floor: i32 = up_count - down_count;
+    let down_count: i64 = input.matches(down_token).count() as i64;
+    let up_count: i64 = input.matches(up_token).count() as i64;
+    let floor: i64 = up_count - down_count;
 
     floor
 }
 
-fn get_basement_position(input: &str) -> i32 {
+fn get_basement_position(input: &str) -> i64 {
     let mut current_floor = 0;
     let basement = -1;
     let mut position = 1;
